videoservice/requirements.txt
asgiref==3.7.2
Authlib==1.2.1
boto3==1.28.3
botocore==1.31.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-cors-headers==4.2.0
django-filter==23.2
django-rest-allauth==1.1
django-rest-framework==0.1.0
django-social-auth==0.7.28
django-storages==1.13.2
djangorestframework==3.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
jmespath==1.0.1
Markdown==3.4.3
oauth2==1.9.0.post1
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-openid==2.2.5
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
zipp==3.16.2
